Events - Work on T3.x and T4
Note: I added some additional debug code, not sure if you want or not.

But on the T3.x I changed the code over to allocate a transfer each time you send an event.  As I found that all of the other USB TX type code in the core3 appears to do this as I believe when you send a packet, when the send is completed that transfer object is returned to the free queue.  The code is also setup to allow up to some N TX packets to be pending (I have it set at 4).

I was thinking of also maybe trying to integrate my other code which does sucessfully send out the Add object, delete object, add directory type events, which works, although you still have to hit F5 to see them.

The issue is that I have additional code added in my version where I basically ask the storage to find the pathname and covert to the node number and then send the event.  Not sure how you want to do such stuff.

But it is in my fork/branch

I updated the test sketch that if you don't have RAM disk it will try to do the add files to SDIO SD... As not really enough memory on T3.6 for decent ram drive