Storage: Add Query functions to storage plus dump command
As I mentioned on forum, added a few query (get functions)  to storage.

First addFileSystem now returns the store number.

Added public method to allow me to query how many filesystems are in the storage.
Added public method to query the name of a store.
Add public method to query the FS for a store (note returns the FS * class.

Added 'd' command to mtp-test which dumps the index list plus dumps the list of storages.