Events: Support adding files plus fix storage change
Here is an updated version of my add Object code that sends events to the host when you do your 'a' command that says here is a new file.

What is different about this version than my previous.  Is if it adds the node to your list of objects, it opens the file in read mode to see if the new object is a directory or not and if not gets the file size.

Then if appropriate it sends the event object to the host.  It is working on both T3.6 and T4.1.

Also fixed the update storage info, so after running the 'a' command and look at the free space, it now appears to update. .

This part appears to be caused by the event that was being sent was not in the list of events we say we do...