WIP - SendObject - finish writes in main loop.
WIP - It will try to read as much as it can into buffers as it is writing.

Once it has read everything into memory it returns to the main loop who responds back to host saying OK... Then in MTPD::loop - it will see that it still has a buffer and issue a write, until it completes the file.

When it completes it will issue a event object prop changed up until then the file shows as 0 length on host...

Next pass, maybe continue up to 3 seconds to complete a write before going to this mode.