WIP
Experiments with:
a) a boolean operator to know if we are connected and have an active session

b) sendObjectInfo, add 3rd parameter by reference for returning the new object id, and have this function return status code.  It checks to see if there is enough free space to hold the file.  Note: this does not compute at all if the actual file will take more space than will actually fit...  now doing a copy of a file that is too big to fit will beep error out.

c) Quick and dirty Format enable.  Right now I allow the code to register a callback with function if the user chooses the format menu item on a storage.  Added callback in example and if they do choose format it will figure out which drive and currenly only with LittleFS will do a quick format.   It then uses sldgehammer to call storage_->ResetIndex() to clear out the index... Could/Should maybe instead simply remofe all of the ojbects that are on the storage that was formatted.