WIP - SendObject - Choose which FS uses new version of send
There is two versions of SendObject here,

The originalish and one that continues to do reads while the Storage is doing writes.
I allow you to specify on an addFileSystem call if you wish to call the new version.

If you use the new version, I allow you to create a secondary buffer and add it to the system, where it uses both buffers...

Still WIP - needs to be cleaned up ... Also what to do with T3.x...