Turn off Send Object Yield code - add timeout code to simple version.
I thought I would try for awhile to figure out ways to recover when a SendObject fails.

I added simple code in the more or less original SendObject that instead of calling pull_packet which waits forever, so timeout the code hangs there.  Instead I unwind that function and function to test to see if packet available and if so then receive that buffer... But in that main test, I have an elapsedMillis which aborts the loop after a fixed time and then bails the copy and now try return of incomplete transfer.

I also put test code to see how long it takes on average to receive the next usb packet.  AVG 0ms  Max 1ms... Timeout is set to half second.  I print error...

On a large copy to propshield it error out, I saw that error message in debug terminal for awhile before I heard the windows error out and the copy dialog disappear .

Still does not recover.  Trying to send the reset event still does not recover....