Update mtp-test.ino
Warning, I think the changes are mainly for changing list of devices
That is the 3,4,5,6 where not all NAND...

So split up for "normal" spi flash and also the NAND winbond.

Warning the changes shows more than it it is, as I think I did the Arduino format on the code which changed some of the spacing

I also add pin 7 to the SPI flash, where I hung off a propshield.

Also turned off the code that created additional buffers as I disabled the code that used this