Format-2Step-IntervalTimer...
First pass - split the format to work in two passes. The first pass is called like before, but in the case of low level format it starts this work on 2nd pass.  The 2nd pass is after it sent response to caller for format.  It starts up an interval timer which does sort of like the main loop code, but on a restricted set of messages.  Any message it does not process it returns bus busy, which if you try to copy a file it will actually put up dialog saying the device is busy...

So partially working.  But I think it is asking for some properties of the formatted device, which I probably return garbage...