Processing command line and detect SD card insert on T4.1
I brought over some of the code I had in earlier branch, so if you just hit enter it tries to give you a list of commands.

Also put in code like FrankB showed that on T4.1 SDCard, if the card is not present at startup, but the user plugs one in, it then adds that store to the storage list.

There are still issues to work out, but if nothing else done, it sends the event for store added. Probably have issues when you have already enum some of the other stores...

But it is a start.