MTP - SD Built in detect SD Card insertion...
Updated it to go ahead and add a missing built in SD Card to the Store to the list at startup time, but I delay adding it until the others if any are added, as I don't want it to try to save away the index files here if we don't have a disk...

Shows up as RED error disk.  If I later insert it and I am not at this folder, it gets the updated data.  If I am there and hit F5 to refresh it shows up correct...

And as the store is in the list, it no longer corrupts the data of other items added that are not roots.