Add support to be able to see the dates and times of file in folders and to preserve them
WIP - Adding support into the underlying projects
starting with FS.h to support the File date/time for Create and Modify, now have ability to let MTP host to ask for these values and display.

This also uncovered that we are not preserving the file dates of the files we create.  So now parse these dates in the SendObjectInfo and then use the new APIs to be able to set the new file to the dates specified.  Which is nice to preserve things like Modify.  Note: MTP does not appear to support Access date, which is probably not worth it anyway